feat: 初始化Spring Boot后端项目基础结构
添加项目配置文件、依赖管理、实体模型、仓储接口和服务层
配置全局异常处理、CORS、Sa-Token认证和WebMVC设置
建立控制器层基础结构并修复实体模型关联关系

+++ springboot-backend/src/main/kotlin/com/yyjy/memo/models/Categories.kt
@@ -1,36 +1,34 @@
 package com.yyjy.memo.models
 
-import org.babyfish.jimmer.sql.Entity
-import org.babyfish.jimmer.sql.Id
-import org.babyfish.jimmer.sql.GeneratedValue
-import org.babyfish.jimmer.sql.Key
-import org.babyfish.jimmer.sql.GenerationType
+import org.babyfish.jimmer.sql.*
 
 /**
-* Entity for table "categories"
-*/
-    @Entity
+ * Entity for table "categories"
+ */
+@Entity
 interface Categories {
 
     @Id
-    @GeneratedValue(            strategy = GenerationType.IDENTITY
-)
-    val id: long
+    @GeneratedValue(
+        strategy = GenerationType.IDENTITY
+    )
+    val id: Int
 
-        /**
-        * 分类名称，如：表白、闲置、树洞
-        */
+
+    /**
+     * 分类名称，如：表白、闲置、树洞
+     */
     @Key
     val name: String
 
-        /**
-        * 分类代表色
-        */
+    /**
+     * 分类代表色
+     */
     val color: String?
 
-        /**
-        * 排序权重
-        */
-    val sortOrder: Integer?
+    /**
+     * 排序权重
+     */
+    val sortOrder: Int?
 }
